Move most of EventSource to shared CoreLib partition
