+++ src/System.Private.CoreLib/shared/Interop/Windows/Advapi32/Interop.EventActivityIdControl.cs
@@ -1,0 +1,24 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System;
+using System.Runtime.InteropServices;
+
+internal partial class Interop
+{
+    internal partial class Advapi32
+    {
+        internal enum ActivityControl : uint
+        {
+            EVENT_ACTIVITY_CTRL_GET_ID = 1,
+            EVENT_ACTIVITY_CTRL_SET_ID = 2,
+            EVENT_ACTIVITY_CTRL_CREATE_ID = 3,
+            EVENT_ACTIVITY_CTRL_GET_SET_ID = 4,
+            EVENT_ACTIVITY_CTRL_CREATE_SET_ID = 5
+        };
+
+        [DllImport(Libraries.Advapi32, ExactSpelling = true)]
+        internal static extern int EventActivityIdControl([In] ActivityControl ControlCode, [In][Out] ref Guid ActivityId);
+    }
+}
