+++ src/System.Private.CoreLib/shared/System/Diagnostics/Tracing/IEventProvider.cs
@@ -17,7 +17,7 @@ namespace System.Diagnostics.Tracing
         // Register an event provider.
         unsafe uint EventRegister(
             EventSource eventSource,
-            UnsafeNativeMethods.ManifestEtw.EtwEnableCallback enableCallback,
+            Interop.Advapi32.EtwEnableCallback enableCallback,
             void* callbackContext,
             ref long registrationHandle);
 
@@ -25,7 +25,7 @@ namespace System.Diagnostics.Tracing
         uint EventUnregister(long registrationHandle);
 
         // Write an event.
-        unsafe int EventWriteTransferWrapper(
+        unsafe int EventWriteTransfer(
             long registrationHandle,
             ref EventDescriptor eventDescriptor,
             IntPtr eventHandle,
@@ -35,7 +35,7 @@ namespace System.Diagnostics.Tracing
             EventProvider.EventData* userData);
 
         // Get or set the per-thread activity ID.
-        int EventActivityIdControl(UnsafeNativeMethods.ManifestEtw.ActivityControl ControlCode, ref Guid ActivityId);
+        int EventActivityIdControl(Interop.Advapi32.ActivityControl ControlCode, ref Guid ActivityId);
 
         // Define an EventPipeEvent handle.
         unsafe IntPtr DefineEventHandle(uint eventID, string eventName, long keywords, uint eventVersion, uint level, byte *pMetadata, uint metadataLength);
